mod: Add new enum in the InfoGame header for each component of the menu

+++ include/gui/InfoGame.hpp
@@ -92,7 +92,7 @@ namespace Gui
     struct Text {
         std::string contentText;
         Pos pos;
-        int fontSize;
+        double fontSize;
         GameColor color;
     };
